+++ ands/ds/MinHeap.py
@@ -6,7 +6,7 @@ Author: Nelson Brochado
 
 Creation: July, 2015
 
-Last update: 17/02/16
+Last update: 19/02/16
 
 A binary min-heap is a data structure similar to a binary tree,
 where the parent nodes are smaller or equal to their children.
@@ -35,7 +35,7 @@ Note that these indexes are for 0-index based lists (or arrays).
 
 - Slides by prof. A. Carzaniga
 
-- Chapter 13 of _Introduction to Algorithms_ (3rd ed.) by CLRS
+- Chapter 13 of [Introduction to Algorithms (3rd ed.)](https://mitpress.mit.edu/books/introduction-algorithms) by CLRS
 """
 
 from ands.ds.Heap import Heap
@@ -50,17 +50,6 @@ class MinHeap(Heap):
     def __init__(self, ls=[]):
         Heap.__init__(self, ls)
 
-    def _build_heap(self):
-        """Creates a min-heap using the list passed to the constructor.
-
-        Note that in a heap A all nodes from A[n/2 + 1] to A[n] are leaf nodes.
-
-        **Time Complexity:** O(n)."""
-        if self.heap:
-            for index in range(len(self.heap) // 2, -1, -1):
-                self.push_down(index)
-        return self.heap
-
     def push_down(self, i: int):
         """'Min-heapify' this min-heap starting from index `i`.
 
@@ -69,9 +58,9 @@ class MinHeap(Heap):
         l = self.left_index(i)
         r = self.right_index(i)
 
-        if l and self.heap[l] < self.heap[m]:
+        if l != -1 and self.heap[l] < self.heap[m]:
             m = l
-        if r and self.heap[r] < self.heap[m]:
+        if r != -1 and self.heap[r] < self.heap[m]:
             m = r
             
         if m != i:
@@ -92,31 +81,12 @@ class MinHeap(Heap):
         c = i  # current index
         p = self.parent_index(i)
         
-        # p could be 0, so we evaluate against not None.
-        if p is not None and self.heap[c] < self.heap[p]:
+        if p != -1 and self.heap[c] < self.heap[p]:
             c = p
 
         if c != i:
             self.swap(c, i)
             self.push_up(c)
-            
-    def add(self, x):
-        """Adds a `x` to this min-heap.
-
-        `x` can either be a key or a `HeapNode` object.
-        If it's a key, an `HeapNode` is first created,
-        whose key and value are equal to `x`.
-         
-        **Time Complexity:** O(log<sub>2</sub> n)."""
-        if x is None:
-            raise ValueError("x cannot be None.")
-        if not isinstance(x, HeapNode):
-            x = HeapNode(x)
-    
-        self.heap.append(x)
-        
-        if self.size() > 1:
-            self.push_up(self.size() - 1)
 
     def find_min(self):
         """Returns (without removing) the smallest element in this min-heap.
@@ -135,63 +105,6 @@ class MinHeap(Heap):
             if not self.is_empty():
                 self.push_down(0)
             return m
-
-    def search(self, x) -> int:
-        """Searches for `x` in this min-heap,
-        and if present, returns its index, otherwise returns -1.
-
-        `x` can either be a key or a `HeapNode` object.
-        If it's a key, an `HeapNode` is first created,
-        whose key and value are equal to `x`.
- 
-        **Time Complexity:** O(n)."""
-        if x is None:
-            raise ValueError("x cannot be None.")
-        if not isinstance(x, HeapNode):
-            x = HeapNode(x)
-
-        for i, node in enumerate(self.heap):
-            if node == x:
-                return i
-        return -1
-
-    def search_by_value(self, val) -> int:
-        """Returns the index of the `HeapNode` object with `value=val`.
-        -1 is returned if no such a `HeapNode` object exists.
-
-        If `val` and the values in this min-heap are not comparable,
-        the behaviour of this method is undefined.
-
-        **Time Complexity:** O(n)."""
-        if val is None:
-            raise ValueError("val cannot be None.")
-        
-        for i, node in enumerate(self.heap):
-            if node.value == val:
-                return i
-        return -1
-    
-    def contains(self, x) -> bool:
-        """Returns True, if `x` is in this min-heap. `False` otherwise.
-
-        `x` can either be a key or a `HeapNode` object.
-        If it's a key, an `HeapNode` is first created,
-        whose key and value are equal to `x`.
-        
-        **Time Complexity:** O(n)."""
-        return self.search(x) != -1    
-
-    def merge(self, other: Heap) -> list:
-        """Merges this min-heap with the `other` heap.
-        
-        Returns the `list` object representing internally the new merged min-heap.
-
-        **Time Complexity:** O(n + m).
-
-        Time complexity analysis based on:
-        [http://stackoverflow.com/a/29197855/3924118](http://stackoverflow.com/a/29197855/3924118)."""
-        self.heap += other.get()
-        return self._build_heap()
 
     def replace(self, i: int, x):
         """Replaces element at index `i` with `x`.
@@ -228,6 +141,7 @@ class MinHeap(Heap):
 
 
 def is_min_heap(h):
+    """Returns `True` if `h` is a valid `MinHeap`. `False` otherwise."""
     if not isinstance(h, MinHeap):
         return False
     if h.heap:
@@ -237,10 +151,10 @@ def is_min_heap(h):
         for i, item in enumerate(h.heap):
             l = h.left_index(i)
             r = h.right_index(i)
-            if r and not l:
+            if r != -1 and l == -1:
                 return False
-            if l and item > h.heap[l]:
+            if l != -1 and item > h.heap[l]:
                 return False
-            if r and item > h.heap[r]:
+            if r != -1 and item > h.heap[r]:
                 return False
     return True  # h is empty