+++ tests/ds/test_MaxHeap.py
@@ -6,6 +6,8 @@ Author: Nelson Brochado
 
 Creation: 17/02/16
 
+Last update: 19/02/16
+
 Tests for the MaxHeap class.
 """
 
@@ -87,8 +89,10 @@ def test_add():
     h.add(12)
     assert not h.is_empty()
     assert h.size() == 1
+    
     assert h.find_max() == HeapNode(12)
-    assert h.heap[0] == h.find_max()    
+    assert h.heap[0] == h.find_max()
+
     assert_is_max_heap(h)
     assert h.remove_max() == HeapNode(12)
     assert h.is_empty()
@@ -123,6 +127,33 @@ def test_add():
     assert h.size() == (s + 1)
     assert h.contains(HeapNode(2048, "2^11"))
     assert_is_max_heap(h)
+
+def test_delete():
+    h = MaxHeap([12, 14, 28, 7, 6, 18])
+
+    def test_bad_index(i):
+        try:
+            h.delete(i)
+            assert False
+        except IndexError:
+            pass
+        except TypeError:
+            pass
+
+    test_bad_index(-1)
+    test_bad_index(None)
+    test_bad_index(6)
+    test_bad_index("0")
+    test_bad_index(3.14159)
+    assert_is_max_heap(h)
+
+    while not h.is_empty():
+        r = randint(0, h.size() - 1)
+        n = h.delete(r)
+        assert n
+        assert_is_max_heap(h)
+
+    assert h.is_empty()
     
 def test_find_max():
     h = MaxHeap([28, 14, 12, 10])    
@@ -347,7 +378,7 @@ def test_parent_index():
     except IndexError:
         pass
 
-    assert not h.parent_index(0)
+    assert h.parent_index(0) == -1
     assert h.parent_index(1) == 0
     assert h.parent_index(2) == 0
 
@@ -366,14 +397,15 @@ def test_grandparent_index():
     except IndexError:
         pass
 
-    assert not h.grandparent_index(0)
-    assert not h.grandparent_index(1)
-    assert not h.grandparent_index(2)
+    assert h.grandparent_index(0) == -1
+    assert h.grandparent_index(1) == -1
+    assert h.grandparent_index(2) == -1
     assert h.grandparent_index(3) == 0
     assert h.grandparent_index(4) == 0
 
 def test_left_index():
     h = MaxHeap([12, 14, 28])
+    
     try:
         h.left_index(-1)
         assert False
@@ -387,8 +419,8 @@ def test_left_index():
         pass
 
     assert h.left_index(0) == 1
-    assert not h.left_index(1)
-    assert not h.left_index(2)
+    assert h.left_index(1) == -1
+    assert h.left_index(2) == -1
 
 def test_right_index():
     h = MaxHeap([12, 14, 28, 6])
@@ -405,9 +437,9 @@ def test_right_index():
         pass
 
     assert h.right_index(0) == 2
-    assert not h.right_index(1)
-    assert not h.right_index(2)
-    assert not h.right_index(3)
+    assert h.right_index(1) == -1
+    assert h.right_index(2) == -1
+    assert h.right_index(3) == -1
     assert h.left_index(1) == 3
 
 def test_has_children():
@@ -550,6 +582,8 @@ def test_is_grandparent():
         assert False
     except IndexError:
         pass
+
+    print(h.is_grandparent(1, 1))
 
     assert not h.is_grandparent(0, 0)
     assert not h.is_grandparent(1, 1)