+++ tests/ds/test_RBT.py
@@ -6,7 +6,7 @@ Author: Nelson Brochado
 
 Creation: 15/02/16
 
-Last update: 17/02/16
+Last update: 19/02/16
 
 Tests for the RBT class.
 """
@@ -14,6 +14,11 @@ Tests for the RBT class.
 from random import randint
 from ands.ds.RBT import RBT, is_rbt, upper_bound_height
 from ands.ds.RBTNode import *
+
+
+from ands.ds.BSTNode import BSTNode
+from ands.ds.BST import BST
+
 
 # ASSERT FUNCTIONS
 
@@ -24,6 +29,7 @@ def assert_rbt_props(t):
         assert not t.root.parent
 
 # TESTS
+
 
 def test_insert_one():
     rbt = RBT()